refactor authentication service and update database schema - change amenityId type from Integer to Long - add login functionality to AuthenticationService - implement JWT token generation for user authentication - update application properties for database connection - create initial application configuration for admin user setup - add role repository for role management

+++ src/main/java/org/example/hotel_service/entities/Amenity.java
@@ -21,7 +21,7 @@ public class Amenity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "amenity_id")
-    Integer amenityId;
+    Long amenityId;
 
     @Column(name = "name", nullable = false, unique = true, length = 120)
     String name;
